+++ Cargo.toml
@@ -9,7 +9,9 @@ anyhow = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["macros", "rt-multi-thread", "signal", "fs", "io-std"] }
+tokio-util = { version = "0.7", features = ["io"] }
+futures-util = { version = "0.3", default-features = false, features = ["std"] }
 once_cell = "1"
 reqwest = { version = "0.12", features = ["rustls-tls", "stream"], default-features = false }
+sync_wrapper = { version = "1", features = ["futures"] }
 beam-lib = { git = "https://github.com/samply/beam", branch = "fix/relative-urls", features = ["http-util", "sockets"] }
-beam-file-lib = { workspace = true }
