+++ beam-file-lib/Cargo.toml
@@ -9,8 +9,9 @@ tokio = { workspace = true }
 beam-lib = { workspace = true }
 serde_json = { workspace = true }
 serde = { workspace = true }
-tokio-util = { version = "0.7", features = ["io"] }
-anyhow = {workspace = true}
-sync_wrapper = { version = "1", features = ["futures"] }
-tracing-subscriber = { workspace = true }
+tokio-util = { workspace = true }
+anyhow = { workspace = true }
+sync_wrapper = { workspace = true }
 tracing = { workspace = true }
+futures-util = { workspace = true }
+once_cell = { workspace = true }
