+++ beam-file-lib/src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod config;
+pub mod beam;
