+++ beam-file/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod config;
-mod server;
+#[cfg(feature = "server")]
+pub mod server;
